+++ srcs/parsing/parsing_utils_2.c
@@ -1,27 +1,22 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   main.c                                             :+:      :+:    :+:   */
+/*   parsing_utils_2.c                                  :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: mpagani <mpagani@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/01/06 11:36:33 by mpagani           #+#    #+#             */
-/*   Updated: 2023/04/01 15:14:33 by mpagani          ###   ########.fr       */
+/*   Created: 2023/04/01 15:03:02 by mpagani           #+#    #+#             */
+/*   Updated: 2023/04/01 15:13:02 by mpagani          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell.h"
 
-int	g_status;
-
-int	main(int argc, char *argv[], char *envp[])
+int	are_quotes(char c)
 {
-	t_minish	*data;
-
-	(void) argv;
-	check_error(argc);
-	data = init_data(envp);
-	setting_prompt(data);
-	exit_clean(data);
-	return (g_status);
+	if (c == '\"')
+		return (1);
+	else if (c == '\'')
+		return (2);
+	return (0);
 }
