+++ srcs/user/here_doc.c
@@ -6,11 +6,11 @@
 /*   By: mpagani <mpagani@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/11 11:40:13 by mpagani           #+#    #+#             */
-/*   Updated: 2023/04/01 10:46:08 by mpagani          ###   ########.fr       */
+/*   Updated: 2023/04/01 15:13:59 by mpagani          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../includes/minishell.h"
+#include "minishell.h"
 
 void	heredoc_write_bible(t_minish *data, int ignore_expansion, char *input)
 {
